backend/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-bstrap-modals==2.1.0
django-cors-headers==3.6.0
django-crum==0.7.9
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-tables2==2.3.3
django-waffle==2.0.0
djangorestframework==3.12.2
edx-django-utils==3.13.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
install==1.3.4
newrelic==5.24.0.153
pbr==5.5.1
phonenumbers==8.12.15
psutil==5.8.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
stevedore==3.3.0
twilio==6.50.1
urllib3==1.26.2
zipp==3.4.0
